requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.21.0
Pillow>=9.0.0
requests>=2.28.0
pydantic>=2.0.0
aiofiles>=23.0.0
